# koicare/koi_swagger.yaml
openapi: 3.0.1
info:
  title: Koi Care API
  description: API documentation for your web application
  version: 1.0.0
servers:
  - url: "http://localhost:80"
    description: Production server
paths:
...
  /pond:
    get:
      summary: Get all ponds
      tags:
        - Pond
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of all ponds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pond"
        "500":
          description: Internal server error
    post:
      summary: Create a new pond
      tags:
        - Pond
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                size:
                  type: number
                  format: float
                depth:
                  type: number
                  format: float
                volume:
                  type: number
                  format: float
                num_of_drains:
                  type: integer
                pump_capacity:
                  type: number
                  format: float
                user_id:
                  type: integer
      responses:
        "201":
          description: Pond created successfully
        "400":
          description: Bad Request
        "500":
          description: Internal server error
  /pond/{id}:
    get:
      summary: Get pond by ID
      tags:
        - Pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the pond
      responses:
        "200":
          description: Pond retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pond"
        "404":
          description: Pond not found
        "500":
          description: Internal server error

    put:
      summary: Update pond by ID
      tags:
        - Pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the pond
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                image:
                  type: string
                size:
                  type: number
                  format: float
                depth:
                  type: number
                  format: float
                volume:
                  type: number
                  format: float
                num_of_drains:
                  type: integer
                pump_capacity:
                  type: number
                  format: float
      responses:
        "200":
          description: Pond updated successfully
        "400":
          description: Bad Request
        "404":
          description: Pond not found
        "500":
          description: Internal server error

    delete:
      summary: Delete pond by ID
      tags:
        - Pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the pond
      responses:
        "200":
          description: Pond deleted successfully
        "404":
          description: Pond not found
        "500":
          description: Internal server error
  /pond/{id}/details:
    get:
      summary: Get pond details by ID
      tags:
        - Pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID of the pond
      responses:
        "200":
          description: Pond details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pond_id:
                    type: integer
                  pond_name:
                    type: string
                  koi_count:
                    type: integer
                  salt_kg_required:
                    type: number
                    format: float
        "404":
          description: Pond not found
        "500":
          description: Internal server error
...
  /pond/{id}/water-info:
    get:
      tags:
        - Pond
      security:
        - bearerAuth: []
      summary: Retrieve water parameters for a specific pond
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the pond
          schema:
            type: integer
      responses:
        '200':
          description: A list of water parameters for the specified pond
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WaterParameter'
        '404':
          description: Pond not found

    post:
      tags:
        - Water Parameters
      security:
        - bearerAuth: []
      summary: Create new water parameters for a specific pond
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the pond
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/WaterParameterInput'
      responses:
        '201':
          description: Water parameters created successfully
        '400':
          description: Invalid input
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pond:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        image:
          type: string
        size:
          type: number
          format: float
        depth:
          type: number
          format: float
        volume:
          type: number
          format: float
        num_of_drains:
          type: integer
        pump_capacity:
          type: number
          format: float
        user_id:
          type: integer
...
    WaterParameter:
      type: object
      properties:
        id:
          type: integer
        temt:
          type: number
        ph:
          type: number
        salinity:
          type: number
        o2:
          type: number
        no2:
          type: number
        no3:
          type: number
        no4:
          type: number
    WaterParameterInput:
      type: object
      properties:
        date:
          type: string
          format: date
        temt:
          type: number
        ph:
          type: number
        salinity:
          type: number
        o2:
          type: number
        no2:
          type: number
        no3:
          type: number
        no4:
          type: number